+++ src/deprecated.jl
@@ -4,3 +4,6 @@
     f,
 )
 @deprecate ExtrinsicEstimation() ExtrinsicEstimation(EfficientEstimator())
+
+Base.@deprecate_binding Symplectic SymplecticMatrices
+#Base.@deprecate_binding SynplecticMatrix SymplecticElement