+++ NEWS.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased] 2023-12-x
+## [0.9.x] – 2024-01-18
 
 ### Added
 
 * added the real symplectic Grassmann manifold `SymplecticGrassmann`
+* Introduce the manifold of `HamiltonianMatrices` and a wrapper for `Hamiltonian` matrices
+* introduce `rand(:HamiltonianMatrices)`
+* Rename (deprecate) `Symplectic` to `SimplecticMatrices` in order to have a `Symplectic` wrapper for such matrices as well
+* Rename (deprecate) `SymplecticMatrix` to `SymplecticElement` to cclarify that it is the special matrix ``J_{2n}`` and not an arbitrary symplectic matrix.
 
 ## [0.9.11] – 2023-12-27